TINY-8279: fixed some circular references
